requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
pydantic
pydantic-settings
python-dotenv
stripe
pyjwt
bcrypt
